fixed deleting past appointments for group's time intervals
