+++ app/background/hourly.js
@@ -22,87 +22,9 @@ const client = require('twilio')(accountSid, authToken);
 // for testing, runs every minute
       // cron.schedule('* * * * *', function() {
         console.log('Scheduled task every day at 55 min past the hour');
-        deletePastAppointments();
         notifyUpcomingAppointments();
-        })
+    })
   } 
-
-  async function deletePastAppointments() {
-    let delDate = new Date().setHours(0,0,0);
-    let delTime = new Date().toLocaleTimeString('it-IT');
-
-    //get all of the appointments for today that start before now
-    await Appointment.findAll({
-      where: { 
-        [Op.or]: [
-          {type: 'Private'},
-          { [Op.and] : [
-              {type: 'Group'},
-              {id : {[Op.in]:  db.sequelize.literal("(SELECT COUNT(spa.id) FROM roles AS sr, personroles as spr, personappointments as spa, appointments a WHERE spr.roleId = sr.id AND spr.personId = spa.personId AND spa.id = a.id AND sr.type = 'Student') = 0")}}
-            ] 
-          }
-        ],
-        status : 'available',
-        date : {[Op.eq]: delDate},
-        startTime : {[Op.lt]: delTime}
-      },
-      include: [{
-        model: Group,
-        as: 'group',
-        required: true
-      }]
-    })
-    .then(appointments => {
-      console.log(delTime+":Checking "+appointments.length+" appointments for deletion or revision");
-      if (appointments.length >0) {
-        // for each appointment check to see if the need to have start time update or be deleted
-        appointments.forEach((appointment) => {
-          let startTime = appointment.startTime.split(":");
-          let hour = parseInt(startTime[0])+1 % 24;
-          if(hour < 10) {
-            hour = "0" + hour
-          }
-          let newStartTime = hour+":"+startTime[1]+":"+startTime[2];
-          // should not try to change time of group appointment, should just delete those
-          if (newStartTime < appointment.endTime && appointment.endTime > delTime && appointment.type === "Private" && appointment.group.allowSplittingAppointments) {
-            appointment.startTime =  newStartTime;
-            let newAppointment = appointment.dataValues
-            let appointmentId = newAppointment.id
-            Appointment.update(newAppointment, {
-              where: { id: appointmentId}
-            })
-            .catch (err => {
-                console.log("Could not update Appointment"+ err);
-            });
-          }
-          else {
-            let id = appointment.id
-            Appointment.destroy({
-              where: { id: id }
-            })
-            .catch (err => {
-              console.log("Could not delete Appointment"+ err);
-            });
-          }
-        })
-      }
-    })
-    .catch(err => {
-      console.log("Could not find past Appointments: "+ err);
-    });
-
-    PersonAppointment.destroy({
-      where: { 
-        appointmentId : null
-      }
-    })
-    .then(num => {
-      console.log( num+' Past person appointments before '+delDate.toString()+' were deleted successfully!');
-    })
-    .catch(err => {
-      console.log("Could not delete past PersonAppointments"+ err);
-    }); 
-  }
 
   async function notifyUpcomingAppointments() {
     let date = new Date().setHours(0,0,0);
@@ -204,6 +126,7 @@ const client = require('twilio')(accountSid, authToken);
         console.log("Could not work 3"+ err);
       });
   }
+  
   function calcTime(time) {
     if(time == null)
     {
